Provide injectables functionality in a mixin.
Allows us to remove it from the base Subsystem class, where it
was causing clutter, as well as introducing an implicit circular
dep on build_graph.
