+++ contrib/scalajs/src/python/pants/contrib/scalajs/subsystems/BUILD
@@ -4,6 +4,7 @@
 python_library(
   dependencies = [
     'contrib/node/src/python/pants/contrib/node/subsystems/resolvers:node_resolver_base',
+    'src/python/pants/build_graph',
     'src/python/pants/option',
     'src/python/pants/subsystem',
     'src/python/pants/util:dirutil',
