+++ src/python/pants/backend/jvm/subsystems/scala_platform.py
@@ -11,6 +11,7 @@ from pants.backend.jvm.subsystems.jvm_tool_mixin import JvmToolMixin
 from pants.backend.jvm.subsystems.zinc_language_mixin import ZincLanguageMixin
 from pants.backend.jvm.targets.jar_library import JarLibrary
 from pants.build_graph.address import Address
+from pants.build_graph.injectables_mixin import InjectablesMixin
 from pants.java.jar.jar_dependency import JarDependency
 from pants.subsystem.subsystem import Subsystem
 
@@ -33,7 +34,7 @@ scala_style_jar = JarDependency('org.scalastyle', 'scalastyle_2.11', '0.8.0')
 
 
 # TODO: Sort out JVM compile config model: https://github.com/pantsbuild/pants/issues/4483.
-class ScalaPlatform(JvmToolMixin, ZincLanguageMixin, Subsystem):
+class ScalaPlatform(JvmToolMixin, ZincLanguageMixin, InjectablesMixin, Subsystem):
   """A scala platform.
 
   :API: public
